requirements.txt
sendgrid
PyGithub
yagmail
jinja2